Remove custom widgets from models and integrate in panel
